feat: improve keyboard handling during tab change
Currently the keyboard is dismissed only when a swipe gesture starts. This commit changes the behaviour so that:

- Keyboard is dismissed if the tab index changes
- When a gesture begines, keyboard is dismissed, but if the index didn't change at end of the gesture, it's restored

This is closer to iOS native behaviour and is based on previous work by [@skevy](https://github.com/skevy) in react-navgation:
https://github.com/react-navigation/react-navigation/pull/3951

The old behaviour is left as the 'on-drag' option for backward compatibility.