fix: Only update state if mounted. (#1081)
### Motivation

This component always logs error in the console when testing.

```
Warning: Can't call setState on a component that is not yet mounted. This is a no-op, but it might indicate a bug in your application. Instead, assign to `this.state` directly or define a `state = {};` class property with the desired state in the Pager component.

Warning: Can't call forceUpdate on a component that is not yet mounted. This is a no-op, but it might indicate a bug in your application. Instead, assign to `this.state` directly or define a `state = {};` class property with the desired state in the Pager component.
```

PR adds some mounted checks.

### Test plan

As long as tests pass and dont log.

Co-authored-by: Miles Johnson <miles.johnson@coinbase.com>